+++ extensions/tidyread---streamline-your-daily-reading/src/utils/util.ts
@@ -77,13 +77,36 @@ export default function formatDate(date: string | number | Date) {
   }
 }
 
-export function retry<T>(fn: () => Promise<T>, retries = 3, delay = 0, err = null): Promise<T> {
-  if (!retries) {
-    return Promise.reject(err);
+export function formatSeconds(seconds: number): string {
+  if (seconds < 60) {
+    return `${seconds}s`;
+  } else {
+    const minutes = Math.floor(seconds / 60);
+    const remainingSeconds = seconds % 60;
+    return `${minutes}min ${remainingSeconds}s`;
   }
+}
+
+export function retry<T>(
+  fn: () => Promise<T>,
+  retries = 3,
+  delay = 0,
+  stopOnError: (error: Error) => boolean = () => false, // 新增参数，判断是否是特定错误
+): Promise<T> {
   return fn().catch((error: any) => {
+    // 检查是否遇到了特定错误
+    if (stopOnError(error)) {
+      return Promise.reject(error);
+    }
+
+    // 当重试次数用完时，返回错误
+    if (retries <= 0) {
+      return Promise.reject(error);
+    }
+
+    // 如果不是特定错误，继续重试
     return new Promise((resolve) => {
-      setTimeout(() => resolve(retry(fn, retries - 1, delay, error)), delay);
+      setTimeout(() => resolve(retry(fn, retries - 1, delay, stopOnError)), delay);
     });
   });
 }