+++ extensions/todoist/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Todoist Changelog
 
+## [Add time to Schedule Task action] - {PR_MERGE_DATE}
+
+It's now possible to specify a time when scheduling a task from the list item.
+
 ## [Change Menu Bar Icon Color in Dark Mode] - 2024-01-22
 
-Changed Menu bar icon color in dark mode to e5e5e5.
+Changed Menu bar icon color in dark mode to #e5e5e5.
 
 ## [Show filters and filter tasks] - 2023-12-18